Add replication pausing, pause menu and relevant behavior (#7)
This does not use OnRep() variables because I had some super weird issues with OnRep() not triggering. I tried debugging for about 10 hours, gave up after. I am aware that this could cause issues for players joining late when the game is currently paused, but I really didn't know why this was happening. I asked in the Unreal discords too, they were puzzled too. :(

History:
* WIP: Add replicated session pausing
Pushed to be able to show to others and get help with a replication issue

* Finish implementation and divert from using OnRep variables to RPCs